Dialog animations added
With this, I think the raw prototype is basically complete. I intend to start a new, private repository for the development of a full-fledged game.

Can't say whether I'll do anything else to maintain/upgrade this side of the project, but I'll leave it up in case anyone wants to take a look at it sometime.